requirements.txt
setuptools
Pillow
ba_colander
cornice
deform
docutils
python_magic
Markdown
moviepy
postmarkup
Pygments
PyYAML
pyramid
pyramid_mailer
pyramid_sqlalchemy
pytz
SQLAlchemy
transaction
WebOb
whoosh
waitress
zope.sqlalchemy
plaster
plaster-pastedeploy
